refactor(ror): consolidate Record of Review components under components/ror
Move all RoR component files (including former atoms) into a single
components/ror/ folder and update import paths.
