+++ client/src/pages/SubmissionDetails.vue
@@ -104,7 +104,7 @@ import AssignedSubmissionUsers from "src/components/AssignedSubmissionUsers.vue"
 import SubmissionAudit from "../components/SubmissionAudit.vue"
 import SubmissionTitle from "src/components/SubmissionTitle.vue"
 import SubmissionExportButton from "src/components/atoms/SubmissionExportButton.vue"
-import RecordOfReviewButton from "src/components/atoms/RecordOfReviewButton.vue"
+import RecordOfReviewButton from "src/components/ror/RecordOfReviewButton.vue"
 import { GET_SUBMISSION } from "src/graphql/queries"
 import { computed, provide, watchEffect, ref } from "vue"
 import { submissionKey } from "src/use/submissionContext"
